added cart per user and guest cart, //todo fix adding items to cart

+++ api/src/main/java/md/master/app/api/ProductDto.java
@@ -1,17 +1,19 @@
 package md.master.app.api;
 
 
+import java.math.BigDecimal;
+
 public class ProductDto {
     private Long id;
     private String title;
-    private int price;
+    private BigDecimal price;
     private String categoryTitle;
 
 
     public ProductDto() {
     }
 
-    public ProductDto(Long id, String title, int price, String categoryTitle) {
+    public ProductDto(Long id, String title, BigDecimal price, String categoryTitle) {
         this.id = id;
         this.title = title;
         this.price = price;
@@ -34,11 +36,11 @@ public class ProductDto {
         this.title = title;
     }
 
-    public int getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(int price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
